chore(ui): remove case-duplicate Badge/Button/Card + normalize imports
