+++ src/components/ui/badge.jsx
@@ -1,34 +1,17 @@
-import * as React from "react"
-import { cva } from "class-variance-authority";
+import * as React from "react";
 
-import { cn } from "@/lib/utils"
-
-const badgeVariants = cva(
-  "inline-flex items-center rounded-md border px-2.5 py-0.5 text-xs font-semibold transition-colors focus:outline-none focus:ring-2 focus:ring-ring focus:ring-offset-2",
-  {
-    variants: {
-      variant: {
-        default:
-          "border-transparent bg-primary text-primary-foreground shadow hover:bg-primary/80",
-        secondary:
-          "border-transparent bg-secondary text-secondary-foreground hover:bg-secondary/80",
-        destructive:
-          "border-transparent bg-destructive text-destructive-foreground shadow hover:bg-destructive/80",
-        outline: "text-foreground",
-      },
-    },
-    defaultVariants: {
-      variant: "default",
-    },
-  }
-)
-
-function Badge({
-  className,
-  variant,
+export function Badge({
+  className = "",
   ...props
 }) {
-  return (<div className={cn(badgeVariants({ variant }), className)} {...props} />);
+  return (
+    <span
+      className={
+        "inline-flex items-center rounded-md border border-bt-border bg-bt-background " +
+        "px-2 py-0.5 text-[12px] font-mono font-medium tracking-[0.05em] uppercase text-bt-ink " +
+        className
+      }
+      {...props}
+    />
+  );
 }
-
-export { Badge, badgeVariants }
